// src/app/models/onboarding/familiares.ts
export class Familiar {
    id: number;
    seguimientoCandidatoId: number;
    parentescoId: number;
    parentescoDescripcion: string;
    nombre: string;
    fechaNacimiento: Date;
    vive: boolean;
    celular: string;
    esBeneficiario: boolean;
    linea: number;
    agregarNuevo: boolean;
    datosModificado: boolean;
}