// src/app/models/onboarding/contactos.ts
export class Contactos {
    id: number;
    seguimientoCandidatoId: number;
    nombre: string;
    relacionId: number;
    relacionDescripcion: string;
    celular: string;
    telefonoTrabajo: string;
    esContactoEmergencia: boolean;
    linea: number;
    agregarNuevo: boolean;
    datosModificado: boolean;
}